ucproxy/requirements.txt
pyOSC==0.3.5b-4996
pyserial==3.2.1
wsgiref==0.1.2
